Moved implicits for the spark package from package.scala to Implicits.scala
